# docs/swagger.yaml
basePath: /
definitions:
...
  dto.NullDateTime:
    properties:
      time:
        type: string
      valid:
        type: boolean
    type: object
...
  echox.PageInfo:
    properties:
      pageNum:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  echox.Response:
    properties:
      code:
        type: string
      data: {}
      message: {}
      page:
        $ref: '#/definitions/echox.PageInfo'
    type: object
...
  system.Notice:
    properties:
      content:
        type: string
      createBy:
        type: integer
      createTime:
        type: string
      id:
        type: integer
      isDeleted:
        type: integer
      level:
        type: string
      publishStatus:
        type: integer
      publishTime:
        $ref: '#/definitions/dto.NullDateTime'
      publisherId:
        type: integer
      revokeTime:
        $ref: '#/definitions/dto.NullDateTime'
      targetType:
        type: integer
      targetUserIds:
        type: string
      title:
        type: string
      type:
        type: integer
      updateBy:
        type: integer
      updateTime:
        type: string
    type: object
  system.NoticeDetailVO:
    properties:
      content:
        type: string
      id:
        type: integer
      level:
        type: string
      publishTime:
        $ref: '#/definitions/dto.NullDateTime'
      publisherId:
        type: integer
      publisherName:
        type: string
      title:
        type: string
      type:
        type: integer
    type: object
  system.NoticeForm:
    properties:
      content:
        type: string
      id:
        type: integer
      level:
        type: string
      targetType:
        type: integer
      targetUserIds:
        items:
          type: string
        type: array
      title:
        maxLength: 50
        type: string
      type:
        type: integer
    required:
    - level
    - targetType
    - title
    - type
    type: object
...
  system.UserNoticePageVO:
    properties:
      id:
        type: integer
      isRead:
        type: integer
      level:
        type: string
      noticeId:
        type: integer
      publishTime:
        $ref: '#/definitions/dto.NullDateTime'
      title:
        type: string
      type:
        type: integer
    type: object
...
host: localhost:2222
info:
  contact:
    email: liusha@email.cn
    name: LiuSha
  description: Echo Admin 后台管理系统 API 文档
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/top-system/light-admin
  title: Echo Admin API
  version: "1.0"
paths:
...
  /api/v1/notices:
    get:
      parameters:
      - description: 标题
        in: query
        name: title
        type: string
      - description: 类型
        in: query
        name: type
        type: integer
      - description: 发布状态
        in: query
        name: publishStatus
        type: integer
      - description: 当前页
        in: query
        name: current
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/echox.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/system.Notice'
                  type: array
              type: object
      summary: 通知公告分页列表
      tags:
      - Notice
    post:
      parameters:
      - description: 通知公告信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/system.NoticeForm'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/echox.Response'
      summary: 新增通知公告
      tags:
      - Notice
  /api/v1/notices/{id}:
    put:
      parameters:
      - description: 通知公告ID
        in: path
        name: id
        required: true
        type: integer
      - description: 通知公告信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/system.NoticeForm'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/echox.Response'
      summary: 修改通知公告
      tags:
      - Notice
  /api/v1/notices/{id}/detail:
    get:
      parameters:
      - description: 通知公告ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/echox.Response'
            - properties:
                data:
                  $ref: '#/definitions/system.NoticeDetailVO'
              type: object
      summary: 获取通知公告详情
      tags:
      - Notice
  /api/v1/notices/{id}/form:
    get:
      parameters:
      - description: 通知公告ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/echox.Response'
            - properties:
                data:
                  $ref: '#/definitions/system.NoticeForm'
              type: object
      summary: 获取通知公告表单数据
      tags:
      - Notice
  /api/v1/notices/{id}/publish:
    put:
      parameters:
      - description: 通知公告ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/echox.Response'
      summary: 发布通知公告
      tags:
      - Notice
  /api/v1/notices/{id}/revoke:
    put:
      parameters:
      - description: 通知公告ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/echox.Response'
      summary: 撤回通知公告
      tags:
      - Notice
  /api/v1/notices/{ids}:
    delete:
      parameters:
      - description: 通知公告ID，多个以英文逗号分割
        in: path
        name: ids
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/echox.Response'
      summary: 删除通知公告
      tags:
      - Notice
  /api/v1/notices/my-page:
    get:
      parameters:
      - description: 标题
        in: query
        name: title
        type: string
      - description: 类型
        in: query
        name: type
        type: integer
      - description: 当前页
        in: query
        name: current
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/echox.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/system.UserNoticePageVO'
                  type: array
              type: object
      summary: 获取我的通知公告分页列表
      tags:
      - Notice
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: 'JWT Authorization header using the Bearer scheme. Example: "Bearer
      {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"